* fix bug: unexpected abort when playing bbs movie. Since fd_set is defined as long type, which used in select() call.
git-svn-id: http://bbs.cs.nthu.edu.tw/svn/maplebbs/trunk@313 bd1103fd-f9b1-db11-a66c-000d613fc1d9

+++ game/tetris.c
@@ -283,7 +283,7 @@ arrived()
 static int
 getkey()
 {
-  int fd;
+  long fd;
   struct timeval tv;
 
   fd = 1;
